Rename ZiplineModule to ZiplineManifest.Module
