+++ zipline-gradle-plugin/src/test/kotlin/app/cash/zipline/gradle/ZiplineGradleDownloaderTest.kt
@@ -16,9 +16,8 @@
 
 package app.cash.zipline.gradle
 
-import app.cash.zipline.loader.ZiplineLoader.Companion.getApplicationManifestFileName
 import app.cash.zipline.loader.ZiplineManifest
-import app.cash.zipline.loader.ZiplineModule
+import app.cash.zipline.loader.internal.getApplicationManifestFileName
 import app.cash.zipline.loader.testing.LoaderTestFixtures
 import java.io.File
 import kotlin.test.assertEquals
@@ -57,7 +56,7 @@ class ZiplineGradleDownloaderTest {
     // Zipline files
     val manifest = ZiplineManifest.create(
       modules = mapOf(
-        "id" to ZiplineModule(
+        "id" to ZiplineManifest.Module(
           url = webServer.url("/latest/app/alpha.zipline").toString(),
           sha256 = testFixtures.alphaSha256,
           dependsOnIds = listOf(),
