+++ zipline-loader/src/commonMain/kotlin/app/cash/zipline/loader/ZiplineLoader.kt
@@ -18,6 +18,9 @@ package app.cash.zipline.loader
 import app.cash.zipline.EventListener
 import app.cash.zipline.Zipline
 import app.cash.zipline.loader.internal.cache.Database
+import app.cash.zipline.loader.internal.cache.SqlDriverFactory
+import app.cash.zipline.loader.internal.cache.ZiplineCache
+import app.cash.zipline.loader.internal.cache.createDatabase
 import app.cash.zipline.loader.internal.fetcher.Fetcher
 import app.cash.zipline.loader.internal.fetcher.FsCachingFetcher
 import app.cash.zipline.loader.internal.fetcher.FsEmbeddedFetcher
@@ -27,9 +30,7 @@ import app.cash.zipline.loader.internal.fetcher.fetch
 import app.cash.zipline.loader.internal.fetcher.fetchManifest
 import app.cash.zipline.loader.internal.fetcher.pin
 import app.cash.zipline.loader.internal.fetcher.unpin
-import app.cash.zipline.loader.internal.cache.SqlDriverFactory
-import app.cash.zipline.loader.internal.cache.ZiplineCache
-import app.cash.zipline.loader.internal.cache.createDatabase
+import app.cash.zipline.loader.internal.getApplicationManifestFileName
 import app.cash.zipline.loader.internal.rebounce
 import app.cash.zipline.loader.internal.receiver.FsSaveReceiver
 import app.cash.zipline.loader.internal.receiver.Receiver
@@ -173,7 +174,7 @@ class ZiplineLoader internal constructor(
    * Creates a Zipline for [applicationName] by fetching the manifest, fetching the modules, loading
    * the code, and running the initializer.
    *
-   * Pass either a [manifestUrl] or a [manifest]. There's no need to pass both.
+   * Pass either a [manifestUrl] or a [providedManifest]. There's no need to pass both.
    */
   internal suspend fun createZiplineAndLoad(
     applicationName: String,
@@ -297,7 +298,7 @@ class ZiplineLoader internal constructor(
   private inner class ModuleJob(
     val applicationName: String,
     val id: String,
-    val module: ZiplineModule,
+    val module: ZiplineManifest.Module,
     val receiver: Receiver,
   ) {
     val upstreams = mutableListOf<Job>()
@@ -343,11 +344,5 @@ class ZiplineLoader internal constructor(
     val manifest = loaded.manifest
     manifestVerifier?.verify(manifestBytes, manifest)
     return loaded
-  }
-
-  companion object {
-    private const val APPLICATION_MANIFEST_FILE_NAME_SUFFIX = "manifest.zipline.json"
-    internal fun getApplicationManifestFileName(applicationName: String) =
-      "$applicationName.$APPLICATION_MANIFEST_FILE_NAME_SUFFIX"
   }
 }
