+++ zipline-loader/src/jvmTest/kotlin/app/cash/zipline/loader/LoaderTester.kt
@@ -17,8 +17,8 @@ package app.cash.zipline.loader
 
 import app.cash.zipline.EventListener
 import app.cash.zipline.Zipline
-import app.cash.zipline.loader.ZiplineLoader.Companion.getApplicationManifestFileName
 import app.cash.zipline.loader.internal.cache.ZiplineCache
+import app.cash.zipline.loader.internal.getApplicationManifestFileName
 import app.cash.zipline.loader.testing.LoaderTestFixtures
 import kotlinx.coroutines.test.TestCoroutineDispatcher
 import okio.FileSystem
