+++ zipline-loader/src/jvmTest/kotlin/app/cash/zipline/loader/ZiplineLoaderTest.kt
@@ -17,8 +17,8 @@ package app.cash.zipline.loader
 
 import app.cash.turbine.test
 import app.cash.zipline.Zipline
-import app.cash.zipline.loader.ZiplineLoader.Companion.getApplicationManifestFileName
 import app.cash.zipline.loader.internal.fetcher.LoadedManifest
+import app.cash.zipline.loader.internal.getApplicationManifestFileName
 import app.cash.zipline.loader.testing.LoaderTestFixtures
 import app.cash.zipline.loader.testing.LoaderTestFixtures.Companion.alphaUrl
 import app.cash.zipline.loader.testing.LoaderTestFixtures.Companion.bravoUrl
